+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- Add native SDK information in the replay option event (#4663)
+
 ## 8.43.1-beta.0
 
 ### Fixes
@@ -32,7 +38,6 @@
 
 ### Improvements
 
-- Add native SDK information in the replay option event (#4663)
 - Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
 - Mask screenshots for errors (#4623)
 - Slightly speed up serializing scope (#4661)